Lower Q so T and Q do not cross
This is a minor adjustment.

Originally, I placed the "Q" on the same baseline as "L" and "T".
This left the right arm of "T" crossing over the loop of "Q".
I liked this, and thought it was consistent
with the original TeX and LaTeX logos.

On further examination, I noticed that none of the letters
in TeX and LaTeX actually cross, they only just touch.

I experimented with shifting "Q" to the right,
so that the arm of "T" would touch rather than cross it,
but I felt there was too much empty space between "Q"
and the bottom serif of "T".

Instead, I decided to lower the "Q", which produces a similar gap
between "T" and "Q" at both top and bottom.